Delete BaseMediationAdapter
